Tie in input value to blend tree animations
Just has the movement animations that the character is doing synced up
with the values being received from the input. Added in some methods
that will set the values of the movement speed and the animator blend
tree dynamically based on the players inputs. Mostly for people playing
with gamepad (currently, until i figure out how to use modifiers for
keyboard and mouse) that will move the character an amount based on how
much they are pressing the joystick.